+++ developer/src/kmc-model/build.sh
@@ -20,6 +20,7 @@ cd "$THIS_SCRIPT_PATH"
 
 builder_describe "Build Keyman kmc Lexical Model Compiler module" \
   "@/common/web/keyman-version" \
+  "@/developer/src/common/web/test-helpers" \
   "configure" \
   "build" \
   "clean" \