+++ developer/src/kmc-package/build.sh
@@ -18,6 +18,7 @@ cd "$THIS_SCRIPT_PATH"
 
 builder_describe "Build Keyman kmc Package Compiler module" \
   "@/common/web/keyman-version" \
+  "@/developer/src/common/web/test-helpers" \
   "configure" \
   "build" \
   "clean" \