+++ developer/src/kmc-package/src/kmp-compiler.ts
@@ -8,6 +8,7 @@ import type { KpsFile, KpsFileContentFile, KpsFileInfo, KpsFileKeyboard, KpsFile
 import type { KmpJsonFile, KmpJsonFileInfo, KmpJsonFileLanguage, KmpJsonFileOptions } from './kmp-json-file.js';
 import { CompilerCallbacks, KvkFile } from '@keymanapp/common-types';
 import { CompilerMessages } from './messages.js';
+import { KMX, KmxFileReader } from '@keymanapp/common-types';
 
 export { type KmpJsonFile } from './kmp-json-file.js';
 
@@ -18,7 +19,8 @@ export default class KmpCompiler {
   constructor(private callbacks: CompilerCallbacks) {
   }
 
-  public transformKpsToKmpObject(kpsString: string, kpsPath: string): KmpJsonFile {
+  // TODO: load the file here via a callback, rather than passing it in as a string
+  public transformKpsToKmpObject(kpsString: string, kpsFilename: string): KmpJsonFile {
 
     // Load the KPS data from XML as JS structured data.
 
@@ -66,9 +68,11 @@ export default class KmpCompiler {
       ['msiOptions', 'msiOptions'],
       ['readMeFile', 'readmeFile']
     ];
-    for (let [src,dst] of keys) {
-      if (kps.options[src]) {
-        kmp.options[dst] = kps.options[src];
+    if(kps.options) {
+      for (let [src,dst] of keys) {
+        if (kps.options[src]) {
+          kmp.options[dst] = kps.options[src];
+        }
       }
     }
 
@@ -78,14 +82,6 @@ export default class KmpCompiler {
 
     if(kps.info) {
       kmp.info = this.kpsInfoToKmpInfo(kps.info);
-    }
-
-    // FollowKeyboardVersion support
-
-    if(kps.options?.followKeyboardVersion !== undefined) {
-      kmp.info.version = {
-        description: this.extractKeyboardVersionFromKmx(kpsPath, kps)
-      };
     }
 
     //
@@ -135,6 +131,17 @@ export default class KmpCompiler {
       kmp.lexicalModels = this.arrayWrap(kps.lexicalModels.lexicalModel).map((model: KpsFileLexicalModel) => {
         return { name:model.name, id:model.iD, languages: this.kpsLanguagesToKmpLanguages(this.arrayWrap(model.languages.language) as KpsFileLanguage[]) }
       });
+    }
+
+    //
+    // FollowKeyboardVersion support
+    //
+
+    if(kps.options?.followKeyboardVersion !== undefined) {
+      kmp.info.version = {
+        description: this.extractKeyboardVersionFromKmx(kpsFilename, kmp)
+      };
+      // TODO: compare the extracted version with other keyboards in the package
     }
 
     //
@@ -195,18 +202,66 @@ export default class KmpCompiler {
     return language.map((element) => { return { name: element._, id: element.$.ID } });
   };
 
-  private extractKeyboardVersionFromKmx(kpsPath: string, kps: KpsFile) {
-    // TODO-LDML: #7340 this is incomplete; we need to read from first .kmx
-    // When we do this, we'll need to update fixtures to have the correct
-    // version string also
-    if(!kps.keyboards || !kps.keyboards.keyboard) {
-      // We don't read from the kps metadata because we want the keyboard
-      // version data here;
-      // TODO: currently model files don't support follow-version?
-      return '0.0.0';
+  private extractKeyboardVersionFromKmx(kpsFilename: string, kmp: KmpJsonFile) {
+    // The DEFAULT_VERSION used to be '1.0', but we now use '0.0' to allow
+    // pre-release 0.x keyboards to be considered later than a keyboard without
+    // any version metadata at all.
+    const DEFAULT_VERSION = '0.0';
+
+    // Note: there is often version metadata in the .kps <Keyboard> element, but
+    // we don't read from the metadata because we want to ensure we have the
+    // most up-to-date keyboard version data here, from the compiled keyboard.
+
+    // Lexical model packages do not allow FollowKeyboardVersion
+    if(kmp.lexicalModels && kmp.lexicalModels.length) {
+      this.callbacks.reportMessage(CompilerMessages.Error_FollowKeyboardVersionNotAllowedForModelPackages());
+      return DEFAULT_VERSION;
     }
-    let k: KpsFileKeyboard[] = this.arrayWrap(kps.keyboards.keyboard);
-    return k?.[0]?.version ?? '0.0.0';
+
+    if(!kmp.keyboards || !kmp.keyboards.length) {
+      this.callbacks.reportMessage(CompilerMessages.Warn_FollowKeyboardVersionButNoKeyboards());
+      return DEFAULT_VERSION;
+    }
+
+    // Reset the keyboard version to the default in the kmp.json metadata, for
+    // warning/failure code paths in this file
+    kmp.keyboards[0].version = DEFAULT_VERSION;
+
+    const file = kmp.files.find(file => path.basename(file.name, '.kmx') == kmp.keyboards[0].id);
+    if(!file) {
+      this.callbacks.reportMessage(CompilerMessages.Error_KeyboardFileNotFound({id:kmp.keyboards[0].id}));
+      return DEFAULT_VERSION;
+    }
+
+    const filename = this.getFullPathToMemberFile(kpsFilename, file.name);
+    if(!fs.existsSync(filename)) {
+      // The zip phase will emit an error later if the file is missing, so
+      // we can just bail cleanly here
+      // console.debug(`The file ${filename} was not found`);
+      return DEFAULT_VERSION;
+    }
+
+    //
+    // load the .kmx and extract the version number
+    //
+    const kmxFileData = fs.readFileSync(filename);
+    const kmxReader: KmxFileReader = new KmxFileReader();
+    const kmx: KMX.KEYBOARD = kmxReader.read(kmxFileData);
+    if(!kmx) {
+      // The file couldn't be read, it might be invalid or locked
+      this.callbacks.reportMessage(CompilerMessages.Error_KeyboardFileNotValid({filename}));
+      return DEFAULT_VERSION;
+    }
+
+    const store = kmx.stores.find(store => store.dwSystemID == KMX.KMXFile.TSS_KEYBOARDVERSION);
+    if(!store) {
+      // We have no version number store, so use default version
+      this.callbacks.reportMessage(CompilerMessages.Warn_KeyboardFileHasNoKeyboardVersion({filename}));
+      return DEFAULT_VERSION;
+    }
+
+    kmp.keyboards[0].version = store.dpString;
+    return store.dpString;
   }
 
   private stripUndefined(o: any) {
@@ -231,8 +286,6 @@ export default class KmpCompiler {
 
     const kmpJsonFileName = 'kmp.json';
 
-    const basePath = path.dirname(kpsFilename);
-
     // Make a copy of kmpJsonData, as we mutate paths for writing
     const data: KmpJsonFile = JSON.parse(JSON.stringify(kmpJsonData));
     if(!data.files) {
@@ -249,20 +302,12 @@ export default class KmpCompiler {
         return;
       }
 
-      if(path.isAbsolute(value.name)) {
-        // absolute paths are not very cross-platform compatible -- we are going to have trouble
-        // with path separators and roots
-        this.callbacks.reportMessage(CompilerMessages.Warn_AbsolutePath({filename: value.name}));
-      } else {
-        // Transform separators to platform separators -- kps files may use
-        // either / or \, although older kps files were always \.
-        if(path.sep == '/') {
-          filename = filename.replace(/\\/g, '/');
-        } else {
-          filename = filename.replace(/\//g, '\\');
-        }
-        filename = path.resolve(basePath, filename);
+      if(path.isAbsolute(filename)) {
+        // absolute paths are not portable to other computers
+        this.callbacks.reportMessage(CompilerMessages.Warn_AbsolutePath({filename: filename}));
       }
+
+      filename = this.getFullPathToMemberFile(kpsFilename, filename);
       const basename = path.basename(filename);
 
       if(!fs.existsSync(filename)) {
@@ -271,18 +316,18 @@ export default class KmpCompiler {
         return;
       }
 
-      let data;
+      let memberFileData;
       try {
-        data = fs.readFileSync(filename);
+        memberFileData = fs.readFileSync(filename);
       } catch(e) {
         this.callbacks.reportMessage(CompilerMessages.Error_FileCouldNotBeRead({filename: filename, e: e}));
         failed = true;
         return;
       }
 
-      this.warnIfKvkFileIsNotBinary(filename, data);
+      this.warnIfKvkFileIsNotBinary(filename, memberFileData);
 
-      zip.file(basename, data);
+      zip.file(basename, memberFileData);
 
       // Remove path data from files before JSON save
       value.name = basename;
@@ -296,6 +341,21 @@ export default class KmpCompiler {
 
     // Generate kmp file
     return zip.generateAsync({type: 'binarystring', compression:'DEFLATE'});
+  }
+
+  private getFullPathToMemberFile(kpsFilename:string, filename: string) {
+    const basePath = path.dirname(kpsFilename);
+    // Transform separators to platform separators -- kps files may use
+    // either / or \, although older kps files were always \.
+    if(path.sep == '/') {
+      filename = filename.replace(/\\/g, '/');
+    } else {
+      filename = filename.replace(/\//g, '\\');
+    }
+    if(!path.isAbsolute(filename)) {
+      filename = path.resolve(basePath, filename);
+    }
+    return filename;
   }
 
   /**